move input file to separate file
