+++ Caprica/main.cpp
@@ -89,7 +89,7 @@ static const std::unordered_set FAKE_SKYRIM_SCRIPTS_SET = {
         "fake://skyrim/DLC1SCWispWallScript.psc",
 };
 
-bool handleImports(const std::vector<conf::ImportFile>& f, caprica::CapricaJobManager* jobManager);
+bool handleImports(const std::vector<ImportFile>& f, caprica::CapricaJobManager* jobManager);
 
 PapyrusCompilationNode* getNode(const PapyrusCompilationNode::NodeType& nodeType,
                                 CapricaJobManager* jobManager,
@@ -107,12 +107,12 @@ PapyrusCompilationNode* getNode(const PapyrusCompilationNode::NodeType& nodeType
                                 time_t lastModTime,
                                 size_t fileSize);
 
-bool addSingleFile(const conf::InputFile& input,
+bool addSingleFile(const InputFile& input,
                    const std::filesystem::path& baseOutputDir,
                    caprica::CapricaJobManager *jobManager,
                    PapyrusCompilationNode::NodeType nodeType);
 
-bool addFilesFromDirectory(const conf::InputFile& input,
+bool addFilesFromDirectory(const InputFile& input,
                            const std::filesystem::path& baseOutputDir,
                            caprica::CapricaJobManager* jobManager,
                            PapyrusCompilationNode::NodeType nodeType,
@@ -287,7 +287,7 @@ PapyrusCompilationNode* getNode(const PapyrusCompilationNode::NodeType& nodeType
   return getNode(nodeType, jobManager, baseOutputDir, curDir, absBaseDir, data.cFileName, lastModTime, fileSize);
 }
 
-bool handleImports(const std::vector<conf::ImportFile>& f, caprica::CapricaJobManager* jobManager) {
+bool handleImports(const std::vector<ImportFile>& f, caprica::CapricaJobManager* jobManager) {
   // Skyrim hacks; we need to import Skyrim's fake scripts into the global namespace first.
   if (conf::Papyrus::game == GameID::Skyrim) {
     caprica::caseless_unordered_identifier_ref_map<PapyrusCompilationNode *> tempMap{};
@@ -320,7 +320,7 @@ bool handleImports(const std::vector<conf::ImportFile>& f, caprica::CapricaJobMa
   return true;
 }
 
-bool addSingleFile(const conf::InputFile& input,
+bool addSingleFile(const InputFile& input,
                    const std::filesystem::path& baseOutputDir,
                    caprica::CapricaJobManager* jobManager,
                    PapyrusCompilationNode::NodeType nodeType) {
